fix radio

+++ Metamapa/M-Agregador-Service/src/main/java/Agregador/DTO/CriterioDTO.java
@@ -18,7 +18,7 @@ public class CriterioDTO {
   private Integer idFuenteDeDatos;
   private Float latitud;
   private Float longitud;
-  private Integer radio;
+  private Float radio;
   private String tipoMultimedia;
   private Boolean inclusion;
